Add files via upload
e-CONIC is a web-based e-learning platform that prioritize simplicity and efficiency in its design. This minimalist approach allows users to focus on the learning material without being overwhelmed by unnecessary distractions. The user interface is clean and uncluttered, with a clear navigation structure that guides users throught the platform’s various features.